helper for converting ENSG to symbols, cbinding symbol and chromosome name for the genes to the input data.frame-like entity

helper for converting ENSG to symbols, cbinding symbol and chromosome name for the genes to the input data.frame-like entity

bind_sym(
  dflike,
  gene_resource = genes_v75,
  ensgvbl = "trait",
  output_symvbl = "symbol",
  output_chromvbl = "gchrom"
)

Arguments

dflike
a data.frame-like entity with a column of ENSG gene identifiers; column name must be the value of bind_sym parameter ensgvbl

gene_resource
a data.frame with columns gene_name and seqnames, defaulting to gwaslake::genes_v75

ensgvbl
character(1) the column of dflike holding the ENSG ids

output_symvbl
character(1) name to be used for symbol in output data.frame, defaults to symbol

output_chromvbl
character(1) name to be used for chromosome in output data.frame, defaults to gchrom

Examples

head(little_eq)

#> Error in gsub(finish, start, ..., fixed = TRUE): zero-length pattern

head(bind_sym(little_eq))

#> Error in gsub(finish, start, ..., fixed = TRUE): zero-length pattern